+++ tests/OsType/ValueBuilders/lsb_release-examples/ubuntu-15.04.php
@@ -1,0 +1,10 @@
+#!/usr/bin/env php
+<?php
+
+echo <<<EOS
+Distributor ID:	Ubuntu
+Description:	Ubuntu 15.04
+Release:	15.04
+Codename:	vivid
+
+EOS;