+++ CHANGELOG.md
@@ -17,3 +17,4 @@
   * Added `GanbaroDigital\OperatingSystem\OsType\Checks\HasEtcRedhatRelease` to check for a Redhat Linux-based distro
   * Added `GanbaroDigital\OperatingSystem\OsType\ValueBuilders\BuildTypeFromEtcIssue` to use `/etc/issue` to determine which Linux distro
   * Added `GanbaroDigital\OperatingSystem\OsType\ValueBuilders\BuildTypeFromEtcRedhatRelease` to use `/etc/redhat-release` to determine which Redhat Linux-based distro we have
+  * Added `GanbaroDigital\OperatingSystem\OsType\ValueBuilders\BuildTypeFromLsbRelease` to use the output of `/usr/bin/lsb_release` to determine which Linux distro we are using